refactor(server): extract writeJSONError to helpers and remove charset suffixes from Content-Type headers
