+++ pkg/constants/constants.go
@@ -39,12 +39,6 @@ const ContentTypeJSON = "application/json"
 // ContentTypePlain is the MIME type for plain text responses.
 const ContentTypePlain = "text/plain"
 
-// ContentTypeJSONCharset is the MIME type for JSON responses with charset.
-const ContentTypeJSONCharset = "application/json; charset=utf-8"
-
-// ContentTypePlainCharset is the MIME type for plain text responses with charset.
-const ContentTypePlainCharset = "text/plain; charset=utf-8"
-
 // AcceptHeaderMaxParts is the maximum number of parts when splitting an Accept header on semicolons.
 const AcceptHeaderMaxParts = 2
