+++ pkg/server/health.go
@@ -1,3 +1,13 @@
+// Package server provides HTTP handlers for the indexer's health, metrics, registration,
+// and schema endpoints.
+//
+// Error response formats differ by audience:
+//   - Host-client endpoints (/api/v1/*) return structured JSON errors via writeJSONError,
+//     using the ErrorResponse envelope {error, message}. These are consumed programmatically
+//     by other indexer clients that need machine-parseable error details.
+//   - User-facing endpoints (/, /health, /registration, /metrics, /snapshots) use plain text
+//     errors via http.Error. These serve browsers and operational dashboards where plain text
+//     is simpler and sufficient.
 package server
 
 import (
